Sort commands by name, add categories
