+++ zabbix_cli/commands/problem.py
@@ -143,6 +143,68 @@ def acknowledge_trigger_last_event(
     )
 
 
+@app.command(name="show_alarms", rich_help_panel=HELP_PANEL)
+def show_alarms(
+    ctx: typer.Context,
+    description: Optional[str] = typer.Option(
+        None,
+        "--description",
+        help="Description of alarm(s) to show.",
+    ),
+    # Could this be a list of priorities in V2?
+    priority: Optional[TriggerPriority] = typer.Option(
+        None,
+        "--priority",
+        help="Priority of alarm(s) to show.",
+    ),
+    hostgroups: Optional[str] = typer.Option(
+        None,
+        "--hostgroup",
+        help="Host group(s) to show alarms for. Comma-separated.",
+    ),
+    unacknowledged: bool = typer.Option(
+        True,
+        "--unack/--ack",
+        help="Show only alarms whose last event is unacknowledged.",
+    ),
+    args: Optional[List[str]] = ARGS_POSITIONAL,
+) -> None:
+    """Show the latest events for the given triggers, hosts, and/or host groups.
+
+    At least one trigger ID, host or host group must be specified.
+    """
+    from zabbix_cli.models import AggregateResult
+
+    if args:
+        if len(args) != 4:
+            exit_err("Invalid number of positional arguments.")
+        description = args[0]
+        priority = TriggerPriority(args[1])
+        hostgroups = args[2]
+        # in V2, "*" was used to indicate "false"
+        if args[3] == "*":
+            unacknowledged = False
+        else:
+            unacknowledged = parse_bool_arg(args[3])
+
+    hostgroups_args = parse_list_arg(hostgroups)
+    hgs = [app.state.client.get_hostgroup(hg) for hg in hostgroups_args]
+    with app.status("Fetching triggers..."):
+        triggers = app.state.client.get_triggers(
+            hostgroups=hgs,
+            description=description,
+            priority=priority,
+            unacknowledged=unacknowledged,
+            select_hosts=True,
+            skip_dependent=True,
+            monitored=True,
+            active=True,
+            expand_description=True,
+            filter={"value": 1},  # why?
+        )
+    render_result(AggregateResult(result=triggers))
+
+
 @app.command(
     name="show_trigger_events",
     rich_help_panel=HELP_PANEL,
@@ -220,65 +282,3 @@ def show_trigger_events(
             limit=limit,
         )
     render_result(AggregateResult(result=events))
-
-
-@app.command(name="show_alarms", rich_help_panel=HELP_PANEL)
-def show_alarms(
-    ctx: typer.Context,
-    description: Optional[str] = typer.Option(
-        None,
-        "--description",
-        help="Description of alarm(s) to show.",
-    ),
-    # Could this be a list of priorities in V2?
-    priority: Optional[TriggerPriority] = typer.Option(
-        None,
-        "--priority",
-        help="Priority of alarm(s) to show.",
-    ),
-    hostgroups: Optional[str] = typer.Option(
-        None,
-        "--hostgroup",
-        help="Host group(s) to show alarms for. Comma-separated.",
-    ),
-    unacknowledged: bool = typer.Option(
-        True,
-        "--unack/--ack",
-        help="Show only alarms whose last event is unacknowledged.",
-    ),
-    args: Optional[List[str]] = ARGS_POSITIONAL,
-) -> None:
-    """Show the latest events for the given triggers, hosts, and/or host groups.
-
-    At least one trigger ID, host or host group must be specified.
-    """
-    from zabbix_cli.models import AggregateResult
-
-    if args:
-        if len(args) != 4:
-            exit_err("Invalid number of positional arguments.")
-        description = args[0]
-        priority = TriggerPriority(args[1])
-        hostgroups = args[2]
-        # in V2, "*" was used to indicate "false"
-        if args[3] == "*":
-            unacknowledged = False
-        else:
-            unacknowledged = parse_bool_arg(args[3])
-
-    hostgroups_args = parse_list_arg(hostgroups)
-    hgs = [app.state.client.get_hostgroup(hg) for hg in hostgroups_args]
-    with app.status("Fetching triggers..."):
-        triggers = app.state.client.get_triggers(
-            hostgroups=hgs,
-            description=description,
-            priority=priority,
-            unacknowledged=unacknowledged,
-            select_hosts=True,
-            skip_dependent=True,
-            monitored=True,
-            active=True,
-            expand_description=True,
-            filter={"value": 1},  # why?
-        )
-    render_result(AggregateResult(result=triggers))
